organize interfaces in different files for better lookup
