+++ src/app/src/definitions/interfaces/alarms_errors.ts
@@ -1,0 +1,10 @@
+export interface AlarmsErrors {
+    type: string,
+    source: string,
+    time: Date,
+    CODE: number,
+    MESSAGE: string,
+    lineNumber: number,
+    line: string,
+    controller: "GRBL" | "grblHAL",
+};
