+++ src/app/src/definitions/interfaces/i18n.ts
@@ -1,0 +1,5 @@
+export interface i18n__Options {
+    context: object,
+    count: number,
+    defaultValue: string,
+};
