+++ src/app/src/definitions/interfaces/jogging.ts
@@ -1,0 +1,17 @@
+export interface JogSpeeds {
+    rapid: {
+        xyStep: string,
+        zStep: string,
+        feedrate: string,
+    },
+    normal: {
+        xyStep: string,
+        zStep: string,
+        feedrate: string,
+    },
+    precise: {
+        xyStep: string,
+        zStep: string,
+        feedrate: string,
+    }
+};
