+++ src/app/src/lib/controller.ts
@@ -22,7 +22,7 @@
  */
 import isElectron from 'is-electron';
 import io, { Socket } from 'socket.io-client';
-import { ControllerListeners } from '../definitions/interfaces';
+import { ControllerListeners } from '../definitions/interfaces/controller';
 
 const ensureArray = (...args: Array<any>) => {
     if (args.length === 0 || args[0] === undefined || args[0] === null) {
